Updates test to deploy to cluster with gitops and validate (#3)

+++ .github/scripts/validate-deploy.sh
@@ -3,6 +3,10 @@
 GIT_REPO=$(cat git_repo)
 GIT_TOKEN=$(cat git_token)
 
+export KUBECONFIG=$(cat .kubeconfig)
+NAMESPACE=$(cat .namespace)
+BRANCH="main"
+
 mkdir -p .testrepo
 
 git clone https://${GIT_TOKEN}@${GIT_REPO} .testrepo
@@ -18,6 +22,7 @@ else
   echo "ArgoCD config found"
 fi
 
+echo "Printing argocd/2-services/active/pact-broker.yaml"
 cat argocd/2-services/active/pact-broker.yaml
 
 if [[ ! -f "payload/2-services/pact-broker/values.yaml" ]]; then
@@ -27,7 +32,39 @@ else
   echo "Application values found"
 fi
 
+echo "Printing payload/2-services/pact-broker/values.yaml"
 cat payload/2-services/pact-broker/values.yaml
+
+count=0
+until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $count -eq 20 ]]; do
+  echo "Waiting for namespace: ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for namespace: ${NAMESPACE}"
+  exit 1
+else
+  echo "Found namespace: ${NAMESPACE}. Sleeping for 30 seconds to wait for everything to settle down"
+  sleep 30
+fi
+
+DEPLOYMENT="pact-broker-${BRANCH}"
+count=0
+until kubectl get deployment "${DEPLOYMENT}" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
+  echo "Waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
+  kubectl get all -n "${NAMESPACE}"
+  exit 1
+fi
+
+kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
 
 cd ..
 rm -rf .testrepo
